requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.2
altair==4.2.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
audio-recorder-streamlit==0.0.10
av==12.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
entrypoints==0.4
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
frozenlist==1.5.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-crc32c==1.6.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
ifaddr==0.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-openai==0.3.0
langchain-text-splitters==0.3.5
langsmith==0.2.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.21.1
numpy==1.26.4
openai==1.59.6
orjson==3.10.14
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.37
streamlit==1.41.1
streamlit-audiorecorder==0.0.6
streamlit-webrtc==0.47.9
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
chromadb==0.4.28
SpeechRecognition==3.10.1
pyaudio==0.2.14
